Processed review comments/sonar issue. Moved command executor to configuration package
