+++ osgp/protocol-adapter-dlms/osgp-protocol-simulator-dlms/simulator/dlms-device-simulator/src/main/resources/application-smr52.properties
@@ -25,8 +25,8 @@ ltediagnostic.capturetime.hundredths=0
 ltediagnostic.capturetime.deviation=-60
 ltediagnostic.capturetime.clockstatus=0
 
-thd.configuration.minDuration.normaltoover=0
-thd.configuration.minDuration.overtonormal=0
-thd.configuration.time.threshold=0
-thd.configuration.value.hysteresis=0
-thd.configuration.value.threshold=0
+thd.configuration.minDuration.normaltoover=900
+thd.configuration.minDuration.overtonormal=600
+thd.configuration.time.threshold=1440
+thd.configuration.value.hysteresis=10
+thd.configuration.value.threshold=86400
